DataAssim.jl

DataAssim.jl

Documentation for DataAssim.jl

Datasets

DataAssim.ETKF — Function.

    Xa,xa = "ensemble_analysis"(Xf,HXf,y,R,H,...)

Computes analysis ensemble Xa based on forecast ensemble Xf
and observations y using various ensemble scheme.
The function name "ensemble_analysis" can be EnSRF, EAKF, ETKF, SEIK, ESTKF, serialEnSRF or EnKF.

Input arguments:

Xf: forecast ensemble (n x N)

HXf: the observation operator applied on the ensemble (product H*Xf)

y: observations (m)

R: observation error covariance (m x m).

H: operator (m x n). Except for the serialEnSRF it is never used and can be empty

Optional keywords arguments:

debug: set to true to enable debugging. Default (false) is no debugging.

tolerance: expected rounding error (default 1e-10) for debugging checks. This is not used if debug is false.

Output arguments:

Xa: the analysis ensemble (n x N)

xa: the analysis ensemble mean (n)

Notations follows: Sangoma D3.1 http://data-assimilation.net/Documents/sangomaDL3.1.pdf

source

DataAssim.local_ETKF — Function.

[Xa,xa] = local_ensemble_analysis(... Xf,H,y,diagR,part,selectObs,method,...)

Computes analysis ensemble Xa based on forecast ensemble Xf using the observation y.

Inputs: Xf: forecast ensemble (n x N) H: observation operator (m x n) y: observation (m x 1) diagR: diagonal of the observation error covariance R (m x 1) part: vector of integer "labels". Every element of the state vector with the same number belong to the same subdomain selectObs: callback routine to select observations with a within a subdomain. As input is takes an integer representing the index of the state vector and returns a vector of weights (m x 1). For example: selectObs(i) = exp(- ((x[i] - xobs[:]).^2 + (y(i) - yobs[:]).^2)/L^2 ); or selectObs(i) = sangoma_compact_locfun(L,... sqrt((x[i] - xobs[:]).^2 + (y[i] - yobs[:]).^2));

where: x and y is the horizontal model grid xobs and yobs are the localtion of the observations L is a correlation length-scale

method: method is one analysis schemes implemented sangoma_ensemble_analysis (except for EnSRF)

Optional inputs: 'display', display: if true, then display progress (false is the default) 'minweight', minweight: analysis is performed using observations for which weights is larger than minweight. (default 1e-8) 'HXf', HXf: if non empty, then it is the product H Xf. In this case, H is not used

Output: Xa: the analysis ensemble (n x N) xa: the analysis ensemble mean (n x 1)

See also: ensemble_analysis, sangoma_compact_locfun

source